+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # MMPReactiveNotification CHANGELOG
 
+## 0.2.0
+
+* Replace singleton with settable service.
+* Settings for remote registration.
+
 ## 0.1.0
 
 * Remote registration signal